fix: correctness burndown — float identity, intervals, load factor (phase 3)
Fixes the runtime-confirmed production bugs from the 2026-06-10 review and
reroutes cmd/validate through production so the harness proves conformance.

- Float tree comparator: treeset/treemap/bag cmp_float.go used an INTRANSITIVE
  "NaN -> raw unsigned bit compare, else natural" construction (NaN's bit
  pattern sorts between negative and positive floats), silently losing tree
  keys when NaN coexisted with negatives. Replaced with the IEEE totalOrder
  sign-flip (== Rust total_cmp), matching the spec.
- Float bit-pattern identity sweep: hash bags, ~28 float-keyed multimaps, and
  the float SentinelHashMap zero-routing now key by math.FloatNbits, so NaN is
  findable and +0/-0 stay distinct (sentinel gains a -0.0 bit sentinel; fixes
  the Size() desync in float bags).
- Float priority queues and Float32/64ArrayList Min/Max/Sort/Distinct/WithoutAll
  now use the total-order comparator / bit-keying (were NaN-unsafe / ±0-collapsing).
- Int64Interval.Get: compute via uint64 mod-2^64 so the result is exact for
  every valid index (no wrap at small indices).
- nextPowerOfTwo: add the 64-bit (>> 32) step; load factor growth condition
  > -> >= across ~120 hash files so it stays strictly below 0.75.
- Regression tests for every fixed family. go build/vet/test clean; validate.sh
  50/50. Reviewed by GPT-5.5 and Fable (both SHIP; an intransitive tree
  comparator they initially missed was found via the arraylist fix and fixed).

+++ interval/int64_interval.go
@@ -84,11 +84,20 @@ func (iv *Int64Interval) Contains(value int64) bool {
 }
 
 // Get returns the element at the given index, or an error if out of bounds.
+//
+// Narrower intervals (int8/16/32) widen into int64 before computing
+// from + step*index so the product never overflows the value width. int64 has
+// no wider native type, so this computes directly at int64 width. Per the spec
+// integer-overflow contract (algorithms.md "Integer overflow contract"), the
+// arithmetic is wrapping two's-complement at the value width — Go's native
+// int64 arithmetic wraps, which is exactly the required semantics. The product
+// is built in uint64 and reinterpreted so the wrap is explicit and free of
+// implementation-defined signed-overflow assumptions.
 func (iv *Int64Interval) Get(index int) (int64, error) {
 	if index < 0 || index >= iv.Size() {
 		return 0, fmt.Errorf("Int64Interval: index out of bounds: %d (size %d)", index, iv.Size())
 	}
-	return int64(int64(iv.from) + int64(iv.step)*int64(index)), nil
+	return int64(uint64(iv.from) + uint64(iv.step)*uint64(int64(index))), nil
 }
 
 func (iv *Int64Interval) absStep() uint64 {
